+++ packages/ngx-flyout/src/lib/flyout-button/flyout-button.module.ts
@@ -1,5 +1,6 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
+import { IconModule } from '@acpaas-ui/ngx-icon';
 
 import {FlyoutModule} from '../flyout/flyout.module';
 
@@ -9,6 +10,7 @@ import {FlyoutButtonComponent} from './components/flyout-button/flyout-button.co
   imports: [
     CommonModule,
     FlyoutModule,
+    IconModule,
   ],
   declarations: [
     FlyoutButtonComponent,