+++ CHANGELOG.md
@@ -9,9 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 - [BREAKING] `icon` New component that completely replaces FontAwesome with Streamline icons
-
-### Changed
 - `avatar` Support for Streamline icons
+- `calendar` Support for Streamline icons
+- `flyout` Support for Streamline icons
 
 ### Removed
 - `cookieconsent` Removed component in favor of our new web component: https://github.com/digipolisantwerp/acpaas-ui_web-components